+++ app/src/androidTest/java/nl/tudelft/cs4160/trustchain_android/GuiEspressoTest/IntegrationGuiTest.java
@@ -7,7 +7,7 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import nl.tudelft.cs4160.trustchain_android.R;
-import nl.tudelft.cs4160.trustchain_android.sharedpreferences.UserNameStorage;
+import nl.tudelft.cs4160.trustchain_android.storage.sharedpreferences.UserNameStorage;
 import nl.tudelft.cs4160.trustchain_android.main.UserConfigurationActivity;
 
 import static android.support.test.InstrumentationRegistry.getInstrumentation;
