--- app/src/main/java/nl/tudelft/cs4160/trustchain_android/bencode/BencodeReadException.java
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2016 Jean Niklas L'orange. All rights reserved.
- *
- * The use and distribution terms for this software are covered by the
- * Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
- * which can be found in the file LICENSE at the root of this distribution.
- *
- * By using this software in any fashion, you are agreeing to be bound by
- * the terms of this license.
- *
- * You must not remove this notice, or any other, from this software.
- */
-package nl.tudelft.cs4160.trustchain_android.bencode;
-
-/**
- * A BencodeReadException is an encoding exception during a Bencode Read.
- *
- * @since 0.1.0
- */
-public class BencodeReadException extends Exception {
-    public BencodeReadException(String message, Object... args) {
-        super(String.format(message, args));
-    }
-}
