+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/inbox/InboxActivity.java
@@ -13,8 +13,8 @@ import java.util.ArrayList;
 import java.util.Collections;
 
 import nl.tudelft.cs4160.trustchain_android.R;
-import nl.tudelft.cs4160.trustchain_android.sharedpreferences.InboxItemStorage;
-import nl.tudelft.cs4160.trustchain_android.peer.Peer;
+import nl.tudelft.cs4160.trustchain_android.storage.sharedpreferences.InboxItemStorage;
+import nl.tudelft.cs4160.trustchain_android.network.peer.Peer;
 
 public class InboxActivity extends AppCompatActivity  {
     public static ArrayList<Peer> peerList;
