+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/inbox/InboxItem.java
@@ -5,7 +5,7 @@ import java.net.InetSocketAddress;
 import java.util.ArrayList;
 import java.util.Arrays;
 
-import nl.tudelft.cs4160.trustchain_android.peer.Peer;
+import nl.tudelft.cs4160.trustchain_android.network.peer.Peer;
 import nl.tudelft.cs4160.trustchain_android.crypto.PublicKeyPair;
 
 public class InboxItem implements Serializable {
