+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/main/UserConfigurationActivity.java
@@ -12,7 +12,7 @@ import android.widget.EditText;
 import android.widget.TextView;
 
 import nl.tudelft.cs4160.trustchain_android.R;
-import nl.tudelft.cs4160.trustchain_android.sharedpreferences.UserNameStorage;
+import nl.tudelft.cs4160.trustchain_android.storage.sharedpreferences.UserNameStorage;
 
 /**
  * The user is able to set his/her own username in this class.
