+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/network/peer/PeerList.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cs4160.trustchain_android.peer;
+package nl.tudelft.cs4160.trustchain_android.network.peer;
 
 import java.util.ArrayList;
