+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/qr/TrustChainBlockFactory.java
@@ -12,7 +12,7 @@ import java.util.Arrays;
 
 import nl.tudelft.cs4160.trustchain_android.block.TrustChainBlockHelper;
 import nl.tudelft.cs4160.trustchain_android.crypto.DualSecret;
-import nl.tudelft.cs4160.trustchain_android.database.TrustChainDBHelper;
+import nl.tudelft.cs4160.trustchain_android.storage.database.TrustChainDBHelper;
 import nl.tudelft.cs4160.trustchain_android.message.MessageProto;
 import nl.tudelft.cs4160.trustchain_android.message.MessageProto.TrustChainBlock.Transaction;
 import nl.tudelft.cs4160.trustchain_android.qr.exception.InvalidDualKeyException;
