+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/storage/database/TrustChainDBHelper.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cs4160.trustchain_android.database;
+package nl.tudelft.cs4160.trustchain_android.storage.database;
 
 import android.content.ContentValues;
 import android.content.Context;
