+++ app/src/test/java/nl/tudelft/cs4160/trustchain_android/network/peer/PeerHandlerTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cs4160.trustchain_android.peer;
+package nl.tudelft.cs4160.trustchain_android.network.peer;
 
 
 import org.junit.After;
