+++ app/src/test/java/nl/tudelft/cs4160/trustchain_android/network/peer/PeerTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cs4160.trustchain_android.peer;
+package nl.tudelft.cs4160.trustchain_android.network.peer;
 
 import junit.framework.TestCase;
