+++ app/src/test/java/nl/tudelft/cs4160/trustchain_android/network/peer/connectionTest/messagesTest/IntroductionResponseTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cs4160.trustchain_android.peer.connectionTest.messagesTest;
+package nl.tudelft.cs4160.trustchain_android.network.peer.connectionTest.messagesTest;
 
 /**
  * Created by michiel on 27-1-2018.
@@ -12,11 +12,11 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
-import nl.tudelft.cs4160.trustchain_android.peer.Peer;
-import nl.tudelft.cs4160.trustchain_android.peer.connection.messages.IntroductionRequest;
-import nl.tudelft.cs4160.trustchain_android.peer.connection.messages.IntroductionResponse;
-import nl.tudelft.cs4160.trustchain_android.peer.connection.messages.Message;
-import nl.tudelft.cs4160.trustchain_android.peer.connection.messages.MessageException;
+import nl.tudelft.cs4160.trustchain_android.network.peer.Peer;
+import nl.tudelft.cs4160.trustchain_android.network.peer.connection.messages.IntroductionRequest;
+import nl.tudelft.cs4160.trustchain_android.network.peer.connection.messages.IntroductionResponse;
+import nl.tudelft.cs4160.trustchain_android.network.peer.connection.messages.Message;
+import nl.tudelft.cs4160.trustchain_android.network.peer.connection.messages.MessageException;
 
 import static org.junit.Assert.assertEquals;
